Iterate strings in JavaScript with a custom iterator
See #69

In this patch I'm trying to see if we can scan through strings in
the library while preserving Unicode semantics if we stay close to
native JavaScript strings. That is, by only accounting for things
like surrogates when we have to.

As is obvious from inspection, this work is incomplete and it was
the attempt that has led me to pursue concverting the intput strings
up-front into an already-split array of Unicode scalar values.