WIP: Implement bisect
It seems like bisect is working except for a quirk where I need to create the `v1` and `v2` arrays with an extra element.

So far I've only seen it completely delete the first text and insert the second so I'm not sure it's working - but it isn't crashing at least.