Enhancement: Render tabs and spaces in pretty HTML previews. (#15)
Previously the `prettyHtml` functions made some changes to diff content
to better represent special characters, but not exhaustively. For
instance, it escapes `<` as `&lt;` so that the HTML preview interprets
the content as plaintext and not as HTML. Similarly, it replaces the
LINE FEED character with `&para;<br>` so that when interpreted by a
browser, the HTML will visually correspond to the newline (whereas HTML
parsers would normalize the whitespace and not show an actual newline).

This patch extends the existing transformation to represent horizontal
tabs with an `&emsp;` EM SPACE and a space character with `&nbsp;` to
further highlight changes to these whitespace characters, rather than
having them blend in via whitespace normalization.

Co-authored-by: Dennis Snell <dmsnell@xkq.io>