Improve bitap algorithm with obvious shortcuts
This patches python3 diff_match_patch.py to take two shortcuts in bitap
score calculation.  First, it does an integer comparison before forcing
a cast to float, and second, if it gets a score of 0.0, it returns it.

All relevant tests pass, as I would expect, because one can demonstrate
that this works just by reading the code carefully.

I know that this patch could be transfered to the C# version as well, I
suspect that all versions have this inefficiency.